fix unit tests, small refactoring
